Alright, i found the bug: I wasnt initializing F_host(1:nb) = 0.
Now I have verified that all the uniform diffusion cases are recreating
the solutions correctly (it all matches what we had before, and the
eyeball norm of the solution looks good; L2 norm is 10^-7 ish).

For the non-uniform diffusion I have interesting resutlts for neumann.
It is a horrible approximation to the solution when using the
SOL_CONSTRAINT. I disabled it and the approximation gets better. I
believe I might need to take the diffusion into consideration for the
SOL_CONSTRAINT, but its not obvious at this time.

Im tired and goign to get a few hours sleep. hopefully I can get this
ironed out before the presentations tonight.