Finally got the neumann BC working well with the discrete RHS. When I switch to analytic RHS things blow up again, but I will do a convergence test to verify that things improve when I increase the # of nodes.
I also updated the Robin boundary condition and got it working to 10^-6
error (relative error on the interior). The boundary is still off, but
when you consider I have the boundary solution equal to 0 analytically,
and the code approximates the robin solution to 10^6, we are within
single precision!

I am commiting and pushing to master and will continue to verify that
the code is working well.

-Evan