Continuing with the paper TOC and draft. We are concentrating on Neumann conditions, so I deleted the Dirichlet figures. I also updated the code and scripts for a plot of the iterative solver residual. Right now we are lookign at r_i/r_0 on a semilogY. This can be easily modified.
I also addressed issues with the relative error calculation which result
in 10^11 when the solution is near 0 and our absolute error is near 0.
For cases which are still bad, the matlab scripts plot the relative
error over the whole domain, then plot the relative error on only the
interior (with boundary nodes showing), and then plot the relative error
on the interior with the relative error on the boundary equal to 0.

-Evan