Fixed a problem with computing the laplacian of MQ when using complex points and complex epsilon.
Also added a test to verify the complex arithmetic which helped me
identify the source of the error in the RBF.

Also updated some of the api to require the problem dimensionality be
specified at class instantiation time because it directly influences the
magnitude of the analytic laplacian of MQ. As we increase the
dimensionality of the MQ we increase the laplacian value, this also
increases the ill-conditioning and instability of our eigenvalues. When
solving the 2D heat equation (e.g., see tests/experiments/original_heat
test) a dimension of 2 allows us to solve the problem. A dimension of 2
for all of the problem EXCEPT the MQ (where we specify dim=3) causes 2
eigenvalues to become unstable in our eigenvalue tests. I need to
improve the tests to understand this better.

TODO: add simple examples for 1D, and 2D using regular grids.
	test convergence
	expand to 2D annulus + test convergence
	continue with paper