Added a new test for solving the parabolic heat equation on a 2D regular grid. I also have notes files added for todo lists as I continue to cleanup the code.
Currently the code sets up and OpenCL kernel inside DerivativeCL
which will allow us to perform the inner product between stencil weights
and stencil solution values. I need to go back and reformat the memory on
the host side so its more direct access on the accelerator.

The code for this new test case does not properly solve a problem yet,
but I am going to push the changes to the master so I avoid getting too
far ahead of master in commits.