Suddenly the binary_cvt_io test case is broken. It has to be something in the memory management down inside the cvt class because Its occuring when I call nested_sphere_cvt->cvt(...) ==> segfault.
I dont want to waste time looking into this, but now is as good a time as any since I need the memory management to work correctly. Ideally the cvt class would inherit the grid_interface so I dont have to do any conversion from CVT to Grid when I want to do a domain decomposiiton.