Got the parallel code working again with a Domain decomposition that inherits from the grid interface. This was important to avoid multiple allocations of memory and to simplify our code so things like the Heat pde only apply to grids/domains (difference between the two is the latter is required to be MPISendable so it supports synchronization during PDE timesteps).
I also am improving the error checking and simplifyin test cases.

My next step is to get the derivative tests working for MPI. after the
transition to the Domain inheriting grid I have a memory segfault on one
of the client procs. Presumably its because a derivative test assumes
nb_stencil == nb_rbfs which is not he case in our decomposed domain