Added support for writing benchmarks to file inside my timer class
Also added new timers to the main routine to benchmark and dump timings
for our code.

Also disabled support for testing eigenvalues in our code until I get
the derivative_tests and derivative class modified. They are currently
operating under the assumption that all boundary stencils appear first
in the stencil_map. Unfortunately, this assumption does not hold when we
split across processes because some of the stencils depend on updates
from neighboring procs and are therefore toward the end of the
stencil_map.