Added a macro to my cmake to allow tests to require PETSC when they build.
Also added the option to specify a working directory for tests so we
can differentiate test cases by subdir.

I have a bug which is preventing me from testing scalability: some of
the code assumes boundary nodes are always first in the stencil_map.
Unfortunately the way I partition stencils, we have some Global Boundary
nodes that appear lower in the map because they are also Ghost Nodes
from neighboring subdomains. I need to account for this.