Added the initial change to var_epsilon in derivative.cpp.
I need to nap and then get back to thinking about the change. Right now
the var_epsilon is not completely variable. For example, when filling a
distance matrix we only use one var_epsilon[irbf] for the entire Mat.
However, in our Ax=b (with A the distance matrix), b is evaluated with
var_epsilon[*].