Next step of the hashing: need to build a list of neighboring cells to search for stencil neighbors.
This is different than gordon and Ian because I need to meet the number
of stencil neighbors requested OR be limited by the radius, OR be
limited by a max number of cell neighbors. The first two require me to
dynamically expand the cell neighbor list until the conditions are met.