Non-linear diffusion is possible. need an exact solution to test it.
Hmm... if time-dependent non-uniform diffusion required careful
derivation of laplacian in 1D, 2D and 3D, its safe to assume that a
spatially dependent diffusion is ALSO going to need to be carefully
derived. I can run a test case ignoring error and see how it looks in
paraview, but I need an exact solution for the paper.