This is a working 2D verification test to prove that the Direct weights are calculated correctly. A few notes:
1) The 13 pt stencil has error too high to continue on the first
iteration on the interior next to the boundary.

2) Symmetric stencils (5, 9) work well.

3) Uniform diffusion only right now. I learned how to generate a general
solution to the problem (in time) and I will make a better test case for
the ellipsoid (fairly certain that the EXACT_SOLUTION is why our code
has been doing so poorly thus far).

4) Code is computing using the DivGrad approach for diffusion. It
appears to work contrary to what Gordon said. Perhaps non-uniform
diffusion will kill it.

5) Im studying Sturm-Liouville Eigenvale Problem to understand how to
write an exact solution for the non-linear diffusion problem.