Alright, lets cover whats going on:
                        Iters to Converge       Residual        ErrorNorm
Viennacl original:          36 / 50 (5)            9e-9            6e-4
CUSP original:              36 / 50 (5)            1e-6            2e-1
VCL Parallel (CUSP):        30 / 50 (5)            9e-9            6e-4

So what I have is an implementation that converges faster with the same
accuracy as VCL for this problem. Only thing is my preconditioner does
not drop the convergence right. I need to figure out what the difference
is between mine and CUSP and fix it ASAP.