Got the NCAR annulus working again. However the solution is poor. I now understand eigenvalues a bit better and I understand how to tune epsilon. What I am seeing is that yes, the system is ill-conditioned. The current parameters for epsilon put the solution near 100% error. HOwever, I need to verify the exact solution AND I need to change the assembly.
When I put eye() on the diagonal of the LHS I am locking eigenvalues
equal to 1. As I found with the stokes, I would get a better solution
with faster convergence if the block corresponding to boundary nodes was
subtracted off the RHS. I will have to do this later. For now, I am
goign to concentrate on writing for the other problems I would like to
use my parallel GMRES on the CVT generated annulus. That ouwld be a
great way to come full circle.